modify dividing function

+++ 20-08-2026/impcalc/Operations/divide.cpp
@@ -4,5 +4,5 @@ double divide(int a, int b) {
         std::cout << "Division by zero is not allowed." << std::endl;
         return 0; // Or handle the error in a way that makes sense for your application
     }
-    return static_cast<double>(a) / static_cast<double>(b);
+    return static_cast<double>(a) / b; 
 }
